clean exists old code
