+++ README.md
@@ -1,8 +1,7 @@
 # copyandpaste
+
 Do not do copy and paste things.
 
-## Post Actions
+## 1. Duplication Options
 
-```bash
-gh search repos util --limit 100 --language=go --json url --jq '.[]|.url' | xargs -I {} gh workflow run .github/workflows/check-any.yaml -F repo_url={}
-```
+## 2. Should not repeat args
