+++ linter_test.go
@@ -10,12 +10,7 @@ func TestAnalyzer(t *testing.T) {
 	testCases := []struct {
 		desc     string
 		settings LinterSetting
-	}{
-		{
-			desc:     "switch",
-			settings: LinterSetting{},
-		},
-	}
+	}{}
 
 	for _, test := range testCases {
 		t.Run(test.desc, func(t *testing.T) {
