+++ repeat_args.go
@@ -1,0 +1,10 @@
+package copyandpaste
+
+import (
+	"go/ast"
+
+	"golang.org/x/tools/go/analysis"
+)
+
+func (a *analyzer) checkRepeatArgs(pass *analysis.Pass, n ast.Node) {
+}
